backend/people-counting/requirements.txt
confluent_kafka==1.5.0
Flask==1.1.2
prometheus-client==0.8.0
python-dotenv==0.15.0
requests==2.24.0
schedule==0.6.0
pymongo==3.11.0
grafana-api==1.0.3
validate_email==1.3
flask_cors==3.0.9
elasticsearch==7.10.0
flask-talisman==0.7.0
python-dateutil==2.8.1  
elasticsearch-dsl==7.3.0
redis==3.5.3
gunicorn==20.0.4
pytz==2020.4
flask-login==0.5.0
bcrypt==3.2.0
sentry-sdk==1.0.0
boto3~=1.17.70
dataclasses~=0.6
Flask-JWT-Extended==3.25.1
PyJwt==1.6.4
sentry-telegram==0.4.0
blinker==1.4
